# Parallelize detensorizing context over batch