**Patch description**
Currently, when generating, we convert the context Tensor to a list one example at a time, which is expensive; this PR parallelizes this operation over the entire batch.

Performance on 1000 generations (`-mf zoo:blender/blender_90M/model -t blended_skill_talk -ne 1000`), on an otherwise unoccupied devfair:

- Original code, trial 1: median elapsed time 783 ms, mean 797 ms
- Original code, trial 2: median 770 ms, mean 785 ms
- This PR, trial 1: median 651 ms, mean 663 ms
- This PR, trial 2: median 665 ms, mean 677 ms

**Testing steps**
CI checks (which seem to cover context blocking)